Basic support for protovalidate
