+++ src/protoj2jsonSchema.ts
@@ -5,7 +5,8 @@ import {Path} from './pathUtils';
 import type {v3} from '@asyncapi/parser/esm/spec-types';
 import {PrimitiveTypes} from './primitive-types';
 import {AsyncAPISchemaDefinition} from '@asyncapi/parser/esm/spec-types/v3';
-import {isOptional as protocGenValidateIsOptional, visit as protocGenValidateVisit} from './protoc-gen-validate';
+import {visit as protocGenValidateVisit} from './protoc-gen-validate';
+import {isOptional as protoValidateIsOptional, visit as protoValidateVisit} from './protovalidate';
 
 const ROOT_FILENAME = 'root';
 const COMMENT_ROOT_NODE = '@RootNode';
@@ -81,6 +82,9 @@ class Proto2JsonSchema {
   private getBundledFileName(filename: string): string | null {
     if (filename === 'validate/validate.proto' || filename === '/validate/validate.proto') {
       return 'validate/validate.proto';
+    }
+    if (filename === 'buf/validate/validate.proto' || filename === '/buf/validate/validate.proto') {
+      return 'buf/validate/validate.proto';
     }
 
     let idx = filename.lastIndexOf('google/protobuf/');
@@ -208,7 +212,7 @@ class Proto2JsonSchema {
   }
 
   private isProto3Required(field: protobuf.Field) {
-    if (protocGenValidateIsOptional(field)) {
+    if (protoValidateIsOptional(field)) {
       return false;
     }
     return (field.options?.proto3_optional !== true && this.isProto3());
@@ -277,6 +281,7 @@ class Proto2JsonSchema {
         }
 
         protocGenValidateVisit(properties[field.name], field);
+        protoValidateVisit(properties[field.name], field);
       } else {
         properties[field.name] = this.compileField(field, item, stack.slice());
       }
@@ -359,6 +364,7 @@ class Proto2JsonSchema {
     this.addDefaultFromCommentAnnotations(obj, field.comment);
     if (!field.repeated) {
       protocGenValidateVisit(obj, field);
+      protoValidateVisit(obj, field);
     }
 
     const desc = this.extractDescription(field.comment);
